Created Router.js and separated routes into single file with helper functions.
Removed sections state from app.vue and am now storing the memory location of the object that holds the refs of each section in the app.
